feat: implement PlaceFacade and update PlaceController for landmark and flower spot search functionality
