+++ pida-core/core-api/src/main/kotlin/com/pida/presentation/v1/flowerspot/FlowerSpotController.kt
@@ -6,10 +6,7 @@ import com.pida.presentation.v1.annotation.ApiV1Controller
 import com.pida.presentation.v1.flowerspot.response.FlowerSpotAllResponse
 import com.pida.presentation.v1.flowerspot.response.FlowerSpotDetailsResponse
 import com.pida.presentation.v1.flowerspot.response.FlowerSpotResponseDto
-import com.pida.presentation.v1.flowerspot.response.FlowerSpotSearchResponse
-import com.pida.presentation.v1.flowerspot.response.PlaceSearchResponse
 import com.pida.support.geo.Region
-import com.pida.user.User
 import io.swagger.v3.oas.annotations.Operation
 import io.swagger.v3.oas.annotations.Parameter
 import io.swagger.v3.oas.annotations.tags.Tag
@@ -52,19 +49,5 @@ class FlowerSpotController(
     ): FlowerSpotDetailsResponse {
         val flowerSpot = flowerSpotFacade.readFlowerSpotDetails(spotId)
         return FlowerSpotDetailsResponse.of(flowerSpot)
-    }
-
-    @Operation(summary = "랜드마크 및 벚꽃길 검색", description = "검색 우선순위에 맞게 랜드마크와 벚꽃길을 검색합니다.")
-    @GetMapping("/flower-spot/search")
-    suspend fun searchFlowerSpot(
-        @RequestParam @Parameter(name = "query", description = "검색 키워드") query: String,
-        @Parameter(hidden = true, required = false) user: User?,
-    ): FlowerSpotSearchResponse {
-        val searchResult = flowerSpotFacade.search(query, user)
-
-        return FlowerSpotSearchResponse.of(
-            landmarks = searchResult.landmarks.map { PlaceSearchResponse.from(it) },
-            flowerSpots = searchResult.flowerSpots.map { PlaceSearchResponse.from(it) },
-        )
     }
 }
